finalize raylib refactor

+++ src/gamestate.h
@@ -4,6 +4,6 @@
 class GameState {
 public:
   virtual ~GameState() = default;
-  virtual void Update(Engine *engine, double frame_delta) = 0;
+  virtual void Update(Engine *engine, float frame_delta) = 0;
   virtual void Draw(Engine *engine) = 0;
 };
