+++ src/menu.cpp
@@ -2,7 +2,7 @@
 #include "game.h"
 #include <raylib.h>
 
-void Menu::Update(Engine *engine, double frame_delta) {
+void Menu::Update(Engine *engine, float frame_delta) {
   if (IsMouseButtonPressed(MOUSE_BUTTON_LEFT)) {
     if (bstart.MouseIsOver()) {
       GameState *game = new Game();
